Decluttering and Bug fixing
