+++ components/SavedLocations/SavedLocationsCSS.css
@@ -1,0 +1,134 @@
+.saved-locations {
+    padding: 0.75rem;
+    margin-top: 1rem;
+}
+
+.saved-locations-grid {
+    display: flex;
+    overflow-x: auto;
+    gap: 0.75rem;
+    padding: 0.25rem 0;
+}
+
+.saved-location-item {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0.5rem 0.875rem;
+    background: rgba(255, 255, 255, 0.08);
+    border-radius: 8px;
+    min-width: 180px;
+    cursor: pointer;
+    border: 1px solid rgba(255, 255, 255, 0.1);
+    transition: all 0.2s ease;
+    gap: 1rem;
+}
+
+.saved-location-item:hover {
+    background: rgba(255, 255, 255, 0.12);
+}
+
+.location-info {
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+}
+
+.location-icon {
+    width: 36px;
+    height: 36px;
+}
+
+.location-icon img {
+    width: 100%;
+    height: 100%;
+    object-fit: contain;
+}
+
+.location-name {
+    font-size: 0.9rem;
+    font-weight: 500;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+}
+
+.location-temp {
+    font-size: 1rem;
+    font-weight: 600;
+    min-width: 50px;
+    text-align: right;
+}
+
+.location-actions {
+    position: relative;
+}
+
+.menu-button {
+    background: transparent;
+    border: none;
+    color: rgba(255, 255, 255, 0.7);
+    cursor: pointer;
+    padding: 0.25rem;
+    border-radius: 4px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    transition: all 0.2s ease;
+}
+
+.menu-button:hover {
+    background-color: rgba(255, 255, 255, 0.1);
+    color: white;
+}
+
+.dropdown-menu {
+    position: absolute;
+    top: 100%;
+    right: 0;
+    background: rgba(16, 17, 36, 0.95);
+    border: 1px solid rgba(255, 255, 255, 0.1);
+    border-radius: 8px;
+    box-shadow: 0 10px 25px rgba(0, 0, 0, 0.2);
+    z-index: 10;
+    width: 150px;
+    overflow: hidden;
+    backdrop-filter: blur(12px);
+    margin-top: 0.5rem;
+    animation: fadeIn 0.2s ease-out;
+}
+
+.dropdown-item {
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    padding: 0.75rem 1rem;
+    color: rgba(255, 255, 255, 0.8);
+    text-decoration: none;
+    width: 100%;
+    text-align: left;
+    background: transparent;
+    border: none;
+    font-size: 0.875rem;
+    transition: all 0.2s ease;
+    cursor: pointer;
+}
+
+.dropdown-item:hover {
+    background: rgba(255, 255, 255, 0.1);
+    color: white;
+}
+
+.dropdown-item svg {
+    color: rgba(255, 255, 255, 0.7);
+}
+
+.dropdown-item:hover svg {
+    color: rgba(255, 0, 0, 0.8);
+}
+
+@media (max-width: 768px) {
+    .saved-location-item {
+        min-width: 150px;
+    }
+}
